Move modify data function into utility file
