+++ src/App.js
@@ -8,6 +8,7 @@ import { fetchAllSongData } from './APICalls';
 import { RiHeartAddLine } from 'react-icons/ri';
 import { MdRemoveCircle } from 'react-icons/md';
 import { FaHeart } from 'react-icons/fa';
+import { modifyData } from './utility';
 
 
 class App extends Component {
@@ -24,7 +25,8 @@ class App extends Component {
     fetchAllSongData()
       .then(data => {
         const [songs, genres] = data;
-        this.modifyData(songs, genres);
+        const modifiedData = modifyData(songs, genres);
+        this.setState({ songBook: modifiedData, renderedSongs: [...this.state.songBook] });
       })
       .catch(err => this.setState({ error: this.handleError(err) }));
   }
@@ -36,22 +38,6 @@ class App extends Component {
       return "Oops, our system seems to be down... how embarassing. Try again in a few moments!"
     }
   }
-
-  modifyData = (songData, genreData) => {
-    const modifiedData = songData.reduce((accu, currentSong, i) => {
-        accu.push(currentSong);
-
-        const matchedItem = genreData.find(genreItem => genreItem.song_id === currentSong.id);
-        let arrayGenres = Object.entries(matchedItem);
-        let foundGenres = arrayGenres.filter(item => item.includes(true));
-
-        accu[i].genres = foundGenres.map(genre => genre[0]);
-        return accu;
-      }, []);
-
-      this.setState({ songBook: modifiedData, renderedSongs: [...this.state.songBook] });
-  }
-
 
   addSong = (id) => {
    const songToAdd = this.state.songBook.find(song => {
